Identify and configuring the device accoding to the software version.
Actually this scripts combine of 3 files (commandsSwitch,commandsRouter,deviceFile). and Autoamate the devices accoding to the information that we instructed. 
Note:- the devices are identifying the device according to the IOS Software version. if we want we can create  a separate file for version. 
commandsSwitch - dedicated configurations for switches
commandsRouter - dedicated configurations for routers
 deviceFile - all devices and those credential information